+++ src/tyr/matrix_serializer.cc
@@ -243,9 +243,15 @@ namespace tyr {
 std::string serializeMatrix(Api& request) {
   double distance_scale = (request.options().units() == Options::miles) ? kMilePerMeter : kKmPerMeter;
 
+  // error if we failed finding any connection
   // dont bother serializing in case of /expansion request
-  if (request.options().action() == Options_Action_expansion)
+  if (std::all_of(request.matrix().times().begin(), request.matrix().times().end(), [](float& time) {
+    return time == kMaxCost;
+  })) {
+    throw valhalla_exception_t(442);
+  } else if (request.options().action() == Options_Action_expansion) {
     return "";
+  }
 
   switch (request.options().format()) {
     case Options_Format_osrm: