+++ CHANGELOG.md
@@ -117,6 +117,7 @@
    * CHANGED: change costmatrix max_distance threshold to a distance threshold instead of duration [#4672](https://github.com/valhalla/valhalla/pull/4672)
    * ADDED: PBF support for expansion [#4614](https://github.com/valhalla/valhalla/pull/4614/)
    * ADDED: elapsed_cost field to map matching json response [#4709](https://github.com/valhalla/valhalla/pull/4709)
+   * ADDED: error if we fail to find any matrix connection [#4718](https://github.com/valhalla/valhalla/pull/4718)
 
 ## Release Date: 2023-05-11 Valhalla 3.4.0
 * **Removed**